Extract `UrlSearch` struct into a submodule.
